buffer: stricter / adjusted stall calculations
- move stall point checks to setextra
- move block vs. buffer size check to ctor
- invert rsp in buffer (to be able to use emp)
- update comments

+++ options.h
@@ -30,7 +30,7 @@ struct options_s {
 	size_t bsiz;
 	size_t rblk, rcnt;
 	size_t wblk, wcnt;
-	size_t rsp, wsp;
+	double rsp, wsp;
 	size_t hpage;
 	int fsync, strict, rline, wline, rcrc, wcrc, loop, cpuR, cpuW;
 	enum mode_t mode;
